pyproject.toml
[tool.poetry]
name = "pochatox/core"
version = "0.1.0"
description = ""
authors = ["EliseyGodX"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
litestar = { extras = ["standart"], version = "^2.14.0" }
uvicorn = "^0.34.0"
redis = "^7.1.1"
pydantic = "^2.12.5"
python-ulid = "^3.1.0"
sqlalchemy = "^2.0.46"
aiosmtplib = "^5.1.0"
kapusta = "^0.1.3"
pyjwt = "^2.11.0"
passlib = "^1.7.4"


[tool.poetry.group.dev.dependencies]
flake8 = "^7.1.1"
flake8-annotations = "^3.1.1"
flake8-bugbear = "^24.12.12"
flake8-docstrings = "^1.7.0"
flake8-isort = "^6.1.1"
flake8-pyproject = "^1.2.3"
mccabe = "^0.7.0"
pycodestyle = "^2.12.1"
pyflakes = "^3.2.0"
unify = "^0.5"
flake8-in-file-ignores = "^0.3.0"
pytest = "^8.3.5"
testcontainers = {extras = ["mailpit"], version = "^4.10.0"}
pytest-asyncio = "^0.26.0"
pytest-mock = "^3.14.0"
deepdiff = "^8.4.2"
wemake-python-styleguide = "1.0"


[tool.flake8]
max-line-length = 88
required-plugins = [
    "flake8-annotations",
    "flake8-bugbear",
    "flake8-docstrings",
    "flake8-isort",
    "flake8-pyproject",
    "mccabe",
    "pycodestyle",
    "pyflakes",
    "flake8-in-file-ignores",
]

ignore = [
    "ANN002",
    "B907",
    "D100",
    "D101",
    "D102",
    "D103",
    "D105",
    "D104",
    "D106",
    "D107",
    "E701",
    "E704",
    "I001",
    "I005",
    "W503",
    "ANN003",
    "ANN101",
    "ANN102",
    "WPS111",
    "WPS201",
    "WPS202",
    "WPS210",
    "WPS214",
    "WPS217",
    "WPS226",
    "WPS229",
    "WPS235",
    "WPS504",
    "WPS602",
    "WPS604",
    "WPS420",
    "WPS430",
    "WPS231",
    "WPS232",
    "WPS238",
    "WPS211",
    "WPS336",

]

select = ["ANN", "B", "C", "D", "E", "F", "I", "T4", "W"]
extend-select = ["B9"]

[tool.pytest.ini_options]
asyncio_default_fixture_loop_scope = "module"
asyncio_default_test_loop_scope = "module"
testpaths = [
    "tests/unit"
]
filterwarnings = [
    "ignore:.*Support for class-based `config` is deprecated.*:DeprecationWarning"
]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
